SPAR-50 Extracted ServiceCardActions
